Move jvmlang agnostic interfaces to jvm.core
Summary:
We would like JavaLibrary to become JvmLibrary at some point in the future. We will try to shrink the java package until it contains only java language specific code. As a first step we'll try to pull the agnostic code into the .core package (or somewhere equally appropriate); this commit begins this process by moving up a set of interfaces that fit the agnostic criterion.