Extract population of JavacOptions from JavaLibraryDescription
Summary:

Shortly we will want to be able to pass a JavacOptions around in
groovy land to support cross compilation. In addition, many classes
currently depend on (and also duplicate an extra piece of)
JavaLibraryDescription's population of JavacOptions from its Arg. In
this change we extract a parent class, JavacArg, and move all
JavacOptions related behaviour to JavacArgInterpreter.

Test-plan: CI