Cross compile groovy and java where appropriate
Summary:

groovyc is able to collaborate with javac in order to compile .java
and .groovy files that express circular dependencies. No comment is
made as to whether this is a good idea, but it is now possible with
buck if one wished to.

By default, the inner javac will be parametrised in the same we it
would be as in a java library rule; these parameters can be overridden
in the same way they can be for a java_library rule.

Test-plan: CI (there are new cases in GroovyLibraryIntegrationTest)